refactor CbrForeignExchangeRateService
